requirements.txt
requests==2.26.0
expects==0.9.0
eggs==0.0.3
psycopg2-binary==2.9.1
dataclasses==0.6
flake8==4.0.1
jupyter==1.0.0
notebook==6.4.6
